+++ packages/smarthr-ui/src/components/Button/index.ts
@@ -1,4 +1,4 @@
 export { Button } from './Button'
-export type { BaseProps } from './types'
+export type { AbstractProps } from './types'
 export { AnchorButton } from './AnchorButton'
 export { UnstyledButton } from './UnstyledButton'