+++ packages/smarthr-ui/src/components/Combobox/MultiCombobox/MultiSelectedItem.tsx
@@ -16,7 +16,7 @@ import { type DecoratorsType, useDecorators } from '../../../hooks/useDecorators
 import { useIntl } from '../../../intl'
 import { UnstyledButton } from '../../Button'
 import { Chip } from '../../Chip'
-import { FaTimesCircleIcon } from '../../Icon'
+import { FaCircleXmarkIcon } from '../../Icon'
 import { Tooltip } from '../../Tooltip'
 import { VisuallyHiddenText } from '../../VisuallyHiddenText'
 
@@ -40,6 +40,7 @@ const classNameGenerator = tv({
     itemLabel: 'smarthr-ui-MultiCombobox-selectedItemLabel',
     deleteButton: [
       'smarthr-ui-MultiCombobox-deleteButton',
+      'shr-relative',
       'shr-group/deleteButton',
       'shr-shrink shr-rounded-full shr-leading-[0] shr-text-black',
       'focus-visible:shr-shadow-[unset]',
@@ -202,7 +203,7 @@ const BaseDestroyButton = <T,>({
       <VisuallyHiddenText id={suffixTextId}>
         {decorated.destroyButtonIconAltSuffix}
       </VisuallyHiddenText>
-      <FaTimesCircleIcon color={disabled ? 'TEXT_DISABLED' : 'inherit'} className={iconStyle} />
+      <FaCircleXmarkIcon color={disabled ? 'TEXT_DISABLED' : 'inherit'} className={iconStyle} />
     </UnstyledButton>
   )
 }