+++ packages/smarthr-ui/src/components/Dialog/StepFormDialog/StepFormDialogContentInner.tsx
@@ -32,7 +32,7 @@ type StepFormHelpers = {
   currentStep: StepItem
 }
 
-export type BaseProps = PropsWithChildren<
+export type AbstractProps = PropsWithChildren<
   DialogHeaderProps &
     DialogBodyProps & {
       /** アクションボタンのラベル */
@@ -54,7 +54,7 @@ export type BaseProps = PropsWithChildren<
     }
 >
 
-export type StepFormDialogContentInnerProps = BaseProps & {
+export type StepFormDialogContentInnerProps = AbstractProps & {
   firstStep: StepItem
   onClickClose: () => void
   responseStatus?: ResponseStatus